requirements.txt
asyncio
aiohttp
websockets
psycopg2-binary
python-dotenv
streamlit
plotly
pandas
